aegis-service/workspace/aegis_workspace/agents/huggingface/requirements.txt
aiohttp
beautifulsoup4
python-dotenv
aegis-sdk